fix stack file syntax and deps
In stack-lts-12.14.yaml, there were dep issues using the new libraries,
which may not be resolvable with those old versions. Didn't check, just
disabled using the libraries there, which will hold us for a couple
years.

(It would probably be easy enough to relax some of the dep versions in
http-client-restricted and git-lfs to the older versions.)